Exposed host in JS WebPack config
By allowing "host" to be specified for the DevServer of a WebPack config, it can be set to "0.0.0.0" to allow access from other computers on the network. See, for example, the discussion here: https://github.com/webpack/webpack-dev-server/issues/147.

Another solution could be to expose the arguments created in KotlinWebpackRunner.kt. However, such a solution would be more invasive.